Reorganize code structure
This splits up the code into functional blocks.
