+++ rustfmt.toml
@@ -1,0 +1,1 @@
+imports_granularity = "Crate"
