+++ src/encoding.rs
@@ -1,0 +1,191 @@
+use std::convert::{TryFrom, TryInto};
+
+use ark_ec::models::TEModelParameters;
+use ark_ed_on_bls12_377::{EdwardsParameters, EdwardsProjective};
+use ark_ff::{Field, One};
+use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
+
+use crate::{constants::TWO, Element, EncodingError, Fq, OnCurve, Sign, SqrtRatioZeta};
+
+#[derive(Copy, Clone, Default, Eq, PartialEq)]
+pub struct Encoding(pub [u8; 32]);
+
+impl std::fmt::Debug for Encoding {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.write_fmt(format_args!(
+            "decaf377::Encoding({})",
+            hex::encode(&self.0[..])
+        ))
+    }
+}
+
+impl Encoding {
+    #[allow(non_snake_case)]
+    pub fn decompress(&self) -> Result<Element, EncodingError> {
+        // This isn't a constant, only because traits don't have const methods
+        // yet and multiplication is only implemented as part of the Mul trait.
+        let D4: Fq = EdwardsParameters::COEFF_D * Fq::from(4u32);
+
+        // 1/2. Reject unless s is canonically encoded and nonnegative.
+        let s = Fq::deserialize(&self.0[..]).map_err(|_| EncodingError::InvalidEncoding)?;
+        if s.is_negative() {
+            return Err(EncodingError::InvalidEncoding);
+        }
+
+        // 3. u_1 <- 1 - s^2
+        let ss = s.square();
+        let u_1 = Fq::one() - ss;
+
+        // 4. u_2 <- u_1^2 - 4d s^2
+        let u_2 = u_1.square() - D4 * ss;
+
+        // 5. sqrt
+        let (was_square, mut v) = Fq::sqrt_ratio_zeta(&Fq::one(), &(u_2 * u_1.square()));
+        if !was_square {
+            return Err(EncodingError::InvalidEncoding);
+        }
+
+        // 6. sign check
+        let two_s_u_1 = *TWO * s * u_1;
+        let check = two_s_u_1 * v;
+        if check.is_negative() {
+            v = -v;
+        }
+
+        // 7. coordinates
+        let x = two_s_u_1 * v.square() * u_2;
+        let y = (Fq::one() + ss) * v * u_1;
+        let z = Fq::one();
+        let t = x * y;
+
+        debug_assert!(
+            EdwardsProjective::new(x, y, t, z).is_on_curve(),
+            "resulting point must be on the curve",
+        );
+
+        Ok(Element {
+            inner: EdwardsProjective::new(x, y, t, z),
+        })
+    }
+}
+
+impl Element {
+    #[allow(non_snake_case)]
+    pub fn compress(&self) -> Encoding {
+        // This isn't a constant, only because traits don't have const methods
+        // yet and subtraction is only implemented as part of the Sub trait.
+        let A_MINUS_D = EdwardsParameters::COEFF_A - EdwardsParameters::COEFF_D;
+        let p = &self.inner;
+
+        // 1.
+        let u_1 = (p.x + p.t) * (p.x - p.t);
+
+        // 2. division by 0 occurs on the identity point, but since
+        // sqrt_ratio_zeta outputs v=0 it computes the right encoding anyway
+        let (_always_square, v) =
+            Fq::sqrt_ratio_zeta(&Fq::one(), &(u_1 * A_MINUS_D * p.x.square()));
+
+        // 3.
+        let u_2 = (v * u_1).abs();
+
+        // 4.
+        let u_3 = u_2 * p.z - p.t;
+
+        // 5.
+        let s = (A_MINUS_D * v * u_3 * p.x).abs();
+
+        // Encode.
+        let mut bytes = [0u8; 32];
+        debug_assert_eq!(s.serialized_size(), 32);
+        s.serialize(&mut bytes[..])
+            .expect("serialization into array should be infallible");
+
+        Encoding(bytes)
+    }
+}
+
+impl From<&Element> for Encoding {
+    fn from(point: &Element) -> Self {
+        point.compress()
+    }
+}
+
+impl From<Element> for Encoding {
+    fn from(point: Element) -> Self {
+        point.compress()
+    }
+}
+
+impl CanonicalSerialize for Encoding {
+    fn serialized_size(&self) -> usize {
+        32
+    }
+
+    fn serialize<W: std::io::Write>(
+        &self,
+        mut writer: W,
+    ) -> Result<(), ark_serialize::SerializationError> {
+        writer.write_all(&self.0[..])?;
+        Ok(())
+    }
+}
+
+impl CanonicalSerialize for Element {
+    fn serialized_size(&self) -> usize {
+        32
+    }
+
+    fn serialize<W: std::io::Write>(
+        &self,
+        writer: W,
+    ) -> Result<(), ark_serialize::SerializationError> {
+        self.compress().serialize(writer)
+    }
+}
+
+impl TryFrom<&[u8]> for Encoding {
+    type Error = EncodingError;
+
+    fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
+        if bytes.len() == 32 {
+            let mut arr = [0u8; 32];
+            arr.copy_from_slice(&bytes[0..32]);
+            Ok(Encoding(arr))
+        } else {
+            Err(EncodingError::InvalidSliceLength)
+        }
+    }
+}
+
+impl TryFrom<&Encoding> for Element {
+    type Error = EncodingError;
+    fn try_from(bytes: &Encoding) -> Result<Self, Self::Error> {
+        bytes.decompress()
+    }
+}
+
+impl TryFrom<Encoding> for Element {
+    type Error = EncodingError;
+    fn try_from(bytes: Encoding) -> Result<Self, Self::Error> {
+        bytes.decompress()
+    }
+}
+
+impl CanonicalDeserialize for Encoding {
+    fn deserialize<R: std::io::Read>(
+        mut reader: R,
+    ) -> Result<Self, ark_serialize::SerializationError> {
+        let mut bytes = [0u8; 32];
+        reader.read_exact(&mut bytes[..])?;
+        Ok(Self(bytes))
+    }
+}
+
+impl CanonicalDeserialize for Element {
+    fn deserialize<R: std::io::Read>(reader: R) -> Result<Self, ark_serialize::SerializationError> {
+        let bytes = Encoding::deserialize(reader)?;
+        bytes
+            .try_into()
+            .map_err(|_| ark_serialize::SerializationError::InvalidData)
+    }
+}
