+++ src/lib.rs
@@ -2,13 +2,21 @@
 //! field](https://penumbra.zone/crypto/primitives/decaf377.html).
 
 mod constants;
+mod element;
+mod elligator;
+mod encoding;
 mod error;
-mod group;
 mod invsqrt;
+mod on_curve;
+mod ops;
 mod sign;
 
+pub use element::Element;
+pub use encoding::Encoding;
 pub use error::EncodingError;
-pub use group::{Element, Encoding};
 
-pub use ark_ed_on_bls12_377::Fq;
-pub use ark_ed_on_bls12_377::Fr;
+pub use ark_ed_on_bls12_377::{Fq, Fr};
+
+use invsqrt::SqrtRatioZeta;
+use on_curve::OnCurve;
+use sign::Sign;
