+++ src/on_curve.rs
@@ -1,0 +1,21 @@
+use ark_ec::models::{twisted_edwards_extended::GroupProjective, TEModelParameters};
+use ark_ff::Field;
+
+pub trait OnCurve {
+    fn is_on_curve(&self) -> bool;
+}
+
+impl<P: TEModelParameters> OnCurve for GroupProjective<P> {
+    #[allow(non_snake_case)]
+    fn is_on_curve(&self) -> bool {
+        let XX = self.x.square();
+        let YY = self.y.square();
+        let ZZ = self.z.square();
+        let TT = self.t.square();
+
+        let on_curve = (YY + P::COEFF_A * XX) == (ZZ + P::COEFF_D * TT);
+        let on_segre_embedding = self.t * self.z == self.x * self.y;
+
+        on_curve && on_segre_embedding
+    }
+}
